requirements.txt
openai==1.93.3
socksio==1.0.0
tiktoken==0.9.0
pandas==2.3.1
openpyxl==3.1.5
python-dotenv==1.1.1